+++ src/hallw/utils/__init__.py
@@ -1,5 +1,15 @@
 from .config_mgr import config
+from .event_bus import Events, emit, subscribe, unsubscribe
 from .hallw_logger import init_logger, logger
 from .prompt_mgr import generateSystemPrompt
 
-__all__ = ["config", "logger", "init_logger", "generateSystemPrompt"]
+__all__ = [
+    "config",
+    "logger",
+    "init_logger",
+    "generateSystemPrompt",
+    "emit",
+    "subscribe",
+    "unsubscribe",
+    "Events",
+]